Completed various tweaks
MTA Viewholder now will automatically switch between a single pair and two pairs of columns, adjusting text scale.
RECY Adapter will now use a new card to replace an existing one of the same type if it exists, and will set MTA to one of the first two cards so it's always at the top of StaggeredLayout columns.
AlertThrower will now accept a Calendar as a parameter to override when the alert should fire.
Calendar display tweaks:
- All day events don't display a time
- Any start or end time outside of today's calendar day will be displayed as "..."
- All LoaderCallbacks for the calendar are entirely contained in the CalendarCallbacks object.